pyproject.toml
[project]
name = "one-pink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colour-science>=0.4.7",
    "matplotlib>=3.10.8",
    "scipy>=1.17.1",
]
